hm_test/requirements.txt
asgiref==3.8.1
attrs==23.2.0
colorama==0.4.6
Django==5.0.3
djangorestframework==3.15.1
drf-spectacular==0.27.2
gunicorn==21.2.0
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
packaging==24.0
pluggy==1.4.0
PyYAML==6.0.1
referencing==0.34.0
rpds-py==0.18.0
sqlparse==0.4.4
tzdata==2024.1
uritemplate==4.1.1
